Reuse the dynamic api prefix resolution for loading terms for public view.
