+++ src/action/AsyncPublicViewActions.ts
@@ -1,17 +1,7 @@
 import { IRI } from "../util/VocabularyUtils";
-import { TermFetchParams, ThunkDispatch } from "../util/Types";
-import { loadTerm, loadTerms } from "./AsyncActions";
+import { ThunkDispatch } from "../util/Types";
+import { loadTerm } from "./AsyncActions";
 import Constants from "../util/Constants";
-
-export function loadPublicTerms(
-  fetchOptions: TermFetchParams<any>,
-  vocabularyIri: IRI
-) {
-  return (dispatch: ThunkDispatch) =>
-    dispatch(
-      loadTerms(fetchOptions, vocabularyIri, Constants.PUBLIC_API_PREFIX)
-    );
-}
 
 export function loadPublicTerm(termFragment: string, vocabularyIri: IRI) {
   return (dispatch: ThunkDispatch) =>
